[Refactor] #200 WalkItem 분리
