+++ DOKI/Presentation/Walk/Model/WalkItem.swift
@@ -1,0 +1,15 @@
+//
+//  WalkItem.swift
+//  DOKI
+//
+//  Created by 이세민 on 1/10/26.
+//
+
+import Foundation
+
+struct WalkItem: Identifiable {
+    let id = UUID()
+    var name: String
+    var isChecked: Bool
+    var isEditing: Bool
+}
